+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Save drawer state in localStorage.
+- Separate landing page and home page routes.
+
+### Fixed
+- Fix auth route not being rendered on SSR.
+
 ### Removed
 - Service worker and workbox plugin on build.