+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Add back button on deck page.
+- Add relevant meta tags.
 
 ### Changed
 - Save drawer state in localStorage.